+++ packages/terra-base/src/Base.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { I18nProvider, i18nLoader } from 'terra-i18n';
+import classNames from 'classnames/bind';
+import { I18nProvider, i18nLoader, injectIntl, intlShape } from 'terra-i18n';
 import './baseStyles';
+import styles from './Base.scss';
+
+const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
@@ -30,13 +34,45 @@ const propTypes = {
    * utilized in this placeholder.
    */
   translationsLoadingPlaceholder: PropTypes.node,
+  /**
+   * The indication as to whether or not the Base component should be sized to match
+   * its parent element's size.
+   */
+  fill: PropTypes.bool,
+  /**
+   * The string selector identifying the element to which Base and its React tree are mounted.
+   */
+  mountSelector: PropTypes.string,
 };
 
 const defaultProps = {
   customMessages: {},
+  mountSelector: '#root',
 };
 
 class Base extends React.Component {
+  static updateMountContainerStyles(mountSelector, fill) {
+    /**
+     * While undefined values pass through querySelector as expected, empty-string values
+     * will throw an exception.
+     */
+    if (!mountSelector || !mountSelector.length) {
+      return;
+    }
+
+    const containerElement = document.querySelector(mountSelector);
+
+    if (!containerElement) {
+      return;
+    }
+
+    if (fill) {
+      containerElement.classList.add(cx('mount-fill'));
+    } else {
+      containerElement.classList.remove(cx('mount-fill'));
+    }
+  }
+
   constructor(props) {
     super(props);
     this.state = {
@@ -55,6 +91,8 @@ class Base extends React.Component {
         console.error(e);
       }
     }
+
+    Base.updateMountContainerStyles(this.props.mountSelector, this.props.fill);
   }
 
   componentWillReceiveProps(nextProps) {
@@ -68,24 +106,36 @@ class Base extends React.Component {
     }
   }
 
+  componentDidUpdate() {
+    Base.updateMountContainerStyles(this.props.mountSelector, this.props.fill);
+  }
+
   render() {
     const {
       children,
       locale,
       customMessages,
       translationsLoadingPlaceholder,
+      fill,
+      mountSelector,
       ...customProps
     } = this.props;
+
+    const baseClassNames = cx([{
+      'base-fill': fill,
+    },
+      customProps.className,
+    ]);
 
     const messages = Object.assign({}, this.state.messages, customMessages);
 
     if (locale === undefined) {
-      return (<div {...customProps}>{children}</div>);
+      return (<div {...customProps} className={baseClassNames}>{children}</div>);
     }
 
-    if (!this.state.areTranslationsLoaded) return <div>{this.props.translationsLoadingPlaceholder}</div>;
+    if (!this.state.areTranslationsLoaded) return <div className={baseClassNames}>{this.props.translationsLoadingPlaceholder}</div>;
     return (
-      <I18nProvider {...customProps} locale={this.state.locale} messages={messages}>
+      <I18nProvider {...customProps} className={baseClassNames} locale={this.state.locale} messages={messages}>
         {children}
       </I18nProvider>
     );
@@ -96,3 +146,4 @@ Base.propTypes = propTypes;
 Base.defaultProps = defaultProps;
 
 export default Base;
+export { injectIntl, intlShape };